# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "x402-tablebook",
    "version": "0.2.0",
    "description": "Restaurant reservations for AI agents, at Chez x402. Availability and booking are paid per call over x402 in USDC; the booking fee is a refundable hold. Lookup and cancellation are free and authenticated by the cancelToken returned at booking time.",
    "x-guidance": "Restaurant reservations for AI agents. Call GET /info (free) for hours, prices and refund policy. Call GET /availability (paid, $0.001) to list open slots — filter with ?date=YYYY-MM-DD&party=N&days=N. Call POST /book (paid, $0.01) with {date, time, party, name} to reserve; the $0.01 is a refundable hold. The booking response returns a cancelToken — store it, it is the only way to look up or cancel the reservation via GET /reservations/{id} and POST /cancel/{id} (both free). Paid routes accept USDC over x402 on either an EVM network or Solana; pick whichever appears in accepts[].",
    "contact": {
      "name": "nirholas",
      "email": "nichxbt@gmail.com",
      "url": "https://github.com/nirholas/x402-tablebook"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4021",
      "description": "This deployment"
    }
  ],
  "paths": {
    "/availability": {
      "get": {
        "operationId": "getAvailability",
        "summary": "Open reservation slots ($0.001 via x402)",
        "description": "Lists bookable slots across the booking window: date, time, seatable party sizes, table types and open table count. All filters are optional; with none supplied the whole window is returned.",
        "tags": [
          "Reservations"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.001000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "Restrict to a single date (YYYY-MM-DD).",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "party",
            "in": "query",
            "required": false,
            "description": "Only slots that can seat this many guests.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "days",
            "in": "query",
            "required": false,
            "description": "How many days ahead to search, from today.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "X-PAYMENT",
            "in": "header",
            "required": false,
            "description": "Base64-encoded signed x402 payment payload. Omit to receive a 402 with PaymentRequirements.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Availability for the requested window",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64-encoded settlement receipt (transaction hash, network, payer)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "restaurant": {
                      "type": "object",
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "description": {
                          "type": "string"
                        },
                        "timezone": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "name",
                        "timezone"
                      ]
                    },
                    "slotMinutes": {
                      "type": "integer"
                    },
                    "seatingMinutes": {
                      "type": "integer"
                    },
                    "refundPolicy": {
                      "type": "object",
                      "properties": {
                        "holdPrice": {
                          "type": "string"
                        },
                        "freeCancellationHours": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "holdPrice",
                        "freeCancellationHours",
                        "description"
                      ]
                    },
                    "generatedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "slots": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date"
                          },
                          "time": {
                            "type": "string",
                            "examples": [
                              "19:00"
                            ]
                          },
                          "partySizes": {
                            "type": "array",
                            "items": {
                              "type": "integer"
                            }
                          },
                          "tableTypes": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "openTables": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "date",
                          "time",
                          "partySizes",
                          "tableTypes",
                          "openTables"
                        ]
                      }
                    }
                  },
                  "required": [
                    "restaurant",
                    "slots",
                    "generatedAt"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment required. Sign one entry from accepts[] and retry with the X-PAYMENT header.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "x402 payment requirements returned with HTTP 402",
                  "properties": {
                    "x402Version": {
                      "type": "integer",
                      "examples": [
                        1
                      ]
                    },
                    "error": {
                      "type": "string"
                    },
                    "accepts": {
                      "type": "array",
                      "description": "One entry per accepted network. Sign whichever your wallet supports.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "scheme": {
                            "type": "string",
                            "examples": [
                              "exact"
                            ]
                          },
                          "network": {
                            "type": "string",
                            "examples": [
                              "base",
                              "solana"
                            ]
                          },
                          "maxAmountRequired": {
                            "type": "string",
                            "description": "Price in USDC atomic units (6 decimals). \"10000\" = $0.01."
                          },
                          "resource": {
                            "type": "string",
                            "format": "uri"
                          },
                          "description": {
                            "type": "string"
                          },
                          "mimeType": {
                            "type": "string"
                          },
                          "payTo": {
                            "type": "string",
                            "description": "Recipient wallet on this network"
                          },
                          "asset": {
                            "type": "string",
                            "description": "USDC contract address or SPL mint"
                          },
                          "maxTimeoutSeconds": {
                            "type": "integer"
                          },
                          "extra": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "scheme",
                          "network",
                          "maxAmountRequired",
                          "resource",
                          "payTo",
                          "asset"
                        ]
                      }
                    }
                  },
                  "required": [
                    "x402Version",
                    "accepts"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/book": {
      "post": {
        "operationId": "bookTable",
        "summary": "Book a table with a refundable hold ($0.01 via x402)",
        "description": "Reserves a table. The fee is a hold, fully refunded on cancellation inside the free window. The response carries the cancelToken — the only credential for looking up or cancelling the reservation — plus a base64 ICS calendar invite and a signed ledger entry.",
        "tags": [
          "Reservations"
        ],
        "x-payment-info": {
          "price": {
            "mode": "fixed",
            "currency": "USD",
            "amount": "0.010000"
          },
          "protocols": [
            {
              "x402": {}
            }
          ]
        },
        "parameters": [
          {
            "name": "X-PAYMENT",
            "in": "header",
            "required": false,
            "description": "Base64-encoded signed x402 payment payload. Omit to receive a 402 with PaymentRequirements.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "YYYY-MM-DD"
                  },
                  "time": {
                    "type": "string",
                    "examples": [
                      "19:00"
                    ],
                    "description": "HH:MM, 24-hour"
                  },
                  "party": {
                    "type": "integer",
                    "minimum": 1,
                    "description": "Number of guests"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Name to hold the table under"
                  },
                  "notes": {
                    "type": "string",
                    "description": "Allergies, occasion, seating preference"
                  },
                  "payerWallet": {
                    "type": "string",
                    "description": "Refund destination. Defaults to the wallet that paid."
                  }
                },
                "required": [
                  "date",
                  "time",
                  "party",
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation confirmed",
            "headers": {
              "X-PAYMENT-RESPONSE": {
                "description": "Base64-encoded settlement receipt (transaction hash, network, payer)",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reservationId": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "confirmed"
                      ]
                    },
                    "restaurant": {
                      "type": "string"
                    },
                    "confirmedTime": {
                      "type": "string",
                      "description": "Human-readable date and time"
                    },
                    "party": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "table": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "type": {
                          "type": "string"
                        },
                        "seats": {
                          "type": "integer"
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "type",
                        "seats"
                      ]
                    },
                    "refundTerms": {
                      "type": "object",
                      "properties": {
                        "holdPrice": {
                          "type": "string"
                        },
                        "freeCancellationHours": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "holdPrice",
                        "freeCancellationHours",
                        "description"
                      ]
                    },
                    "cancelToken": {
                      "type": "string",
                      "description": "Store this — required to look up or cancel the reservation."
                    },
                    "cancelEndpoint": {
                      "type": "string"
                    },
                    "ledgerEntry": {
                      "type": "object",
                      "properties": {
                        "entryId": {
                          "type": "string"
                        },
                        "reservationId": {
                          "type": "string"
                        },
                        "kind": {
                          "type": "string",
                          "enum": [
                            "hold",
                            "refund",
                            "forfeit"
                          ]
                        },
                        "amount": {
                          "type": "string"
                        },
                        "wallet": {
                          "type": "string"
                        },
                        "reason": {
                          "type": "string"
                        },
                        "at": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "entryId",
                        "reservationId",
                        "kind",
                        "amount",
                        "reason",
                        "at"
                      ]
                    },
                    "ics": {
                      "type": "string",
                      "description": "base64-encoded RFC 5545 calendar invite"
                    },
                    "signature": {
                      "type": "string",
                      "description": "HMAC-SHA256 over canonical JSON"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  },
                  "required": [
                    "reservationId",
                    "status",
                    "confirmedTime",
                    "party",
                    "name",
                    "table",
                    "cancelToken",
                    "createdAt"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error code"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable explanation"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          },
          "402": {
            "description": "Payment required. Sign one entry from accepts[] and retry with the X-PAYMENT header.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "x402 payment requirements returned with HTTP 402",
                  "properties": {
                    "x402Version": {
                      "type": "integer",
                      "examples": [
                        1
                      ]
                    },
                    "error": {
                      "type": "string"
                    },
                    "accepts": {
                      "type": "array",
                      "description": "One entry per accepted network. Sign whichever your wallet supports.",
                      "items": {
                        "type": "object",
                        "properties": {
                          "scheme": {
                            "type": "string",
                            "examples": [
                              "exact"
                            ]
                          },
                          "network": {
                            "type": "string",
                            "examples": [
                              "base",
                              "solana"
                            ]
                          },
                          "maxAmountRequired": {
                            "type": "string",
                            "description": "Price in USDC atomic units (6 decimals). \"10000\" = $0.01."
                          },
                          "resource": {
                            "type": "string",
                            "format": "uri"
                          },
                          "description": {
                            "type": "string"
                          },
                          "mimeType": {
                            "type": "string"
                          },
                          "payTo": {
                            "type": "string",
                            "description": "Recipient wallet on this network"
                          },
                          "asset": {
                            "type": "string",
                            "description": "USDC contract address or SPL mint"
                          },
                          "maxTimeoutSeconds": {
                            "type": "integer"
                          },
                          "extra": {
                            "type": "object"
                          }
                        },
                        "required": [
                          "scheme",
                          "network",
                          "maxAmountRequired",
                          "resource",
                          "payTo",
                          "asset"
                        ]
                      }
                    }
                  },
                  "required": [
                    "x402Version",
                    "accepts"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Slot unavailable or outside opening hours",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Machine-readable error code"
                    },
                    "message": {
                      "type": "string",
                      "description": "Human-readable explanation"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}